fix: refactor routes
