+++ server/app.js
@@ -5,7 +5,7 @@ const connectDB = require("./configs/db.config");
 var cors = require("cors");
 const corsOptions = require("./configs/corsOptions");
 const requireAuth = require("./middlewares/requireAuth");
-const authRoutes = require("./routes/auth");
+const authRoutes = require("./routes/user/auth");
 const vendorRoutes = require("./routes/vendor/routes");
 const adminRoutes = require("./routes/admin/routes");
 const userRoutes = require("./routes/user/routes");
